To test this change: have a test that should fail fast and run it. You will see that it takes `3m` to fail with timeout. In my case I killed my local docker registry and run a test that uses it:

Before:
```sh
go clean -testcache && go test --tags=integration -v -run  TestStoreImpl
=== RUN   TestStoreImpl
=== PAUSE TestStoreImpl
=== CONT  TestStoreImpl
--- FAIL: TestStoreImpl (180.00s)
    store_test.go:162: micro run failure, output: EOF
        
    store_test.go:162: micro run failure, output: EOF
        
    store_test.go:162: micro run failure, output: EOF
        
    test_framework.go:535: micro run failure, output: EOF
    store_test.go:148: store_test.go:148, TestStoreImpl (failed after 3m0.000111713s)
FAIL
exit status 1
FAIL	github.com/micro/micro/v3/test	180.005s

```

After:
```sh
$ go clean -testcache && go test --tags=integration -v -run  TestStoreImpl
=== RUN   TestStoreImpl
=== PAUSE TestStoreImpl
=== CONT  TestStoreImpl
--- FAIL: TestStoreImpl (9.48s)
    store_test.go:162: micro run failure, output: EOF
        
    store_test.go:162: micro run failure, output: EOF
        
    store_test.go:162: micro run failure, output: EOF
        
    store_test.go:148: micro run failure, output: EOF
FAIL
exit status 1
FAIL	github.com/micro/micro/v3/test	9.484s
```

By using `watch -n1 docker status` it's clear that the container long exists before the end of the timeout. I suspect this is because `t.Fatal` terminates the function where it's being called, hence `done <- true` never runs.

Edit:

Indeed `Fatal` calls `FailNow` which
`FailNow marks the function as having failed and stops its execution
// by calling runtime.Goexit (which then runs all deferred calls in the`